2005-04-09	Ralf Corsepius <ralf.corsepius@rtems.org>
	* libmisc/monitor/mon-symbols.c: size_t.

+++ cpukit/libmisc/monitor/mon-symbols.c
@@ -70,7 +70,7 @@ rtems_symbol_create(
     uint32_t       value
     )
 {
-    int symbol_length;
+    size_t symbol_length;
     size_t newsize;
     rtems_symbol_t *sp;
